+++ CHANGES.markdown
@@ -7,6 +7,8 @@
 * Drop dependency on compojure except for examples.
 * #201 Add support for using a java.net.URI instance to specify
   a Location for `moved` handlers
+* Posting to an existing resource checks for conflicts.
+
 
 # New in 0.14.1