+++ CHANGES.markdown
@@ -10,6 +10,7 @@
 * Posting to an existing resource checks for conflicts.
 * Add `:post-enacted?`, `:put-enacted?` and `:patch-enacted?`
   which return status 202 accepted if true.
+* Add leiningen alias `graph` to generate `trace.svg`
 
 # New in 0.14.1